backend/org-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
strawberry-graphql==0.220.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg[binary]==3.1.18
